backend/requirements.txt
Flask==3.0.2
Flask-Cors
Flask-Caching
requests==2.28.1
python-dotenv==0.21.0
gunicorn==20.1.0
click
pnpm
waitress